move shuffledsongs to playback_state
